water_app/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
marshmallow==3.21.3
mongoengine==0.28.2
packaging==24.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
